+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/lang/zh-CN/
 
 ## [Unreleased]
 
+## [0.11.0] - 2026-06-09
+
 ### Added
 
 - 集成 `nonebot-plugin-permission`，支持用 `command.wordcloud.*` 权限资源管理词云命令。
@@ -334,7 +336,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/lang/zh-CN/
 
 - 可以使用的版本。
 
-[Unreleased]: https://github.com/he0119/nonebot-plugin-wordcloud/compare/v0.10.0...HEAD
+[Unreleased]: https://github.com/he0119/nonebot-plugin-wordcloud/compare/v0.11.0...HEAD
+[0.11.0]: https://github.com/he0119/nonebot-plugin-wordcloud/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/he0119/nonebot-plugin-wordcloud/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/he0119/nonebot-plugin-wordcloud/compare/v0.8.1...v0.9.0
 [0.8.1]: https://github.com/he0119/nonebot-plugin-wordcloud/compare/v0.8.0...v0.8.1